requirements.txt
asyncio
pyserial-asyncio
pyudev
